Hack/Fix hinter with empty token
When the hinter token is '.', the default textual hints get shown along
with the tern results. This is because the list has nothing to filter
on. If we just remove the textual-hints behavior, we end up with the
filter-list deactivating before tern can respond with the completion
results and nothing gets displayed. If we could control the order
the behaviors are processed during object/raise-reduce, this wouldn't
be an issue, but it doesn't appear to be an easy fix with the give
platform api.